// src/modules/v1/scoring/enums/key.enum.ts
export enum I18NEnum {
    EN = "en",
    VI = "vi"
}

export enum KeyForKeyEnum {
    INPUT = "INPUT",
    PICTURE_CONTENT = "PICTURE_CONTENT",
    REQUIRED_WORD = "REQUIRED_WORD",
    OPTION_WORDS = "OPTION_WORDS",
    ANSWER = "ANSWER",
    REQUIRED_WORDS = "REQUIRED_WORDS",
    PARAGRAPH_TOPIC = "PARAGRAPH_TOPIC",
    MIA_ANSWER = "MIA_ANSWER",
    LEARN_SUMMARY = "LEARN_SUMMARY",
    YES = "YES",
    NO = "NO"
}

export enum KeyUsereENEnum {
    INPUT = "Input",
    PICTURE_CONTENT = "Picture content",
    REQUIRED_WORD = "Require word",
    OPTION_WORDS = "Option words",
    ANSWER = "User answers",
    REQUIRED_WORDS = "Required words",
    PARAGRAPH_TOPIC = "Title content",
    MIA_ANSWER = "Exam answers",
    LEARN_SUMMARY = "Learner answers",
    YES = "YES",
    NO = "NO"
}

export enum KeyUserVIEnum {
    INPUT = "Đầu vào",
    PICTURE_CONTENT = "Nội dung hình ảnh",
    REQUIRED_WORD = "Từ bắt buộc",
    OPTION_WORDS = "Các từ tùy chọn",
    ANSWER = "Câu trả lời của người dùng",
    REQUIRED_WORDS = "Các từ bắt buộc",
    PARAGRAPH_TOPIC = "Nội dung tiêu đề",
    MIA_ANSWER = "Đáp án đề thi",
    LEARN_SUMMARY = "Câu trả lời của người học",
    YES = "CÓ",
    NO = "KHÔNG"
}